+++ packages/layout/jest.config.ts
@@ -1,10 +1,10 @@
 export default {
-  displayName: 'sidebar',
+  displayName: 'layout',
   preset: '../../jest.preset.js',
   transform: {
     '^(?!.*\\.(js|jsx|ts|tsx|css|json)$)': '@nx/react/plugins/jest',
     '^.+\\.[tj]sx?$': ['babel-jest', { presets: ['@nx/react/babel'] }]
   },
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory: '../../coverage/packages/sidebar'
+  coverageDirectory: '../../coverage/packages/layout'
 }
