+++ packages/layout/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@midas-ds/sidebar",
+  "name": "@midas-ds/layout",
   "version": "0.0.1",
   "main": "./index.js",
   "types": "./index.d.ts",
