+++ packages/layout/project.json
@@ -1,7 +1,7 @@
 {
-  "name": "sidebar",
+  "name": "layout",
   "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "packages/sidebar/src",
+  "sourceRoot": "packages/layout/src",
   "projectType": "library",
   "tags": [],
   "targets": {
@@ -12,7 +12,7 @@
       "executor": "@nx/eslint:lint",
       "outputs": ["{options.outputFile}"],
       "options": {
-        "lintFilePatterns": ["packages/select/**/*.{ts,tsx,js,jsx}"]
+        "lintFilePatterns": ["packages/layout/**/*.{ts,tsx,js,jsx}"]
       }
     },
     "build": {
@@ -20,7 +20,7 @@
       "outputs": ["{options.outputPath}"],
       "defaultConfiguration": "production",
       "options": {
-        "outputPath": "dist/packages/sidebar"
+        "outputPath": "dist/packages/layout"
       },
       "configurations": {
         "production": {
@@ -36,7 +36,7 @@
       "executor": "@nx/jest:jest",
       "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
       "options": {
-        "jestConfig": "packages/sidebar/jest.config.ts",
+        "jestConfig": "packages/layout/jest.config.ts",
         "passWithNoTests": true
       },
       "configurations": {
